+++ knowledge-base-public/package.json
@@ -35,6 +35,8 @@
   },
   "devDependencies": {
     "autoprefixer": "10.4.13",
+    "eslint": "^8.44.0",
+    "eslint-plugin-react": "^7.32.2",
     "postcss": "8.4.18",
     "prettier": "2.7.1",
     "prettier-plugin-tailwindcss": "0.1.13",
