+++ knowledge-base-public/src/pages/_document.js
@@ -11,7 +11,6 @@ class MyDocument extends Document {
     return (
       <Html>
         <Head>
-          <link href="https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap" rel="stylesheet" />
           <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet" />
           <script defer data-domain="support.snu.gouv.fr" src="https://plausible.io/js/plausible.js"></script>
         </Head>
